Add counter.js file that contains commentclaer
